Improve parsing for offset and /proc/self/maps
  * Remove the use of `line_no` which was erroneous: when main was
    mapped to memory, its address does not necessary start at 0.
    For exemple, if the second sector is mapped with the execution
    rights (hence, being the second entry in `/proc/self/maps`), the
    base address needs to be set.
  * Compute the offset using standard string function, instead of
    iterating manually over the buffers.
  * Remove unnecessary tests. For exemple in `strlen(module)>0` which is
    implied by the boolean return value.

+++ src/utils.hxx
@@ -5,5 +5,5 @@
 #pragma once
 
 bool parse_proc_self_maps_entry (const char *entry,
-	size_t *start, size_t *end, unsigned lenpermissions, char *permissions,
-	size_t *offset, unsigned lenmodule, char *module);
+	size_t *start, size_t *end, size_t lenpermissions, char *permissions,
+	size_t *offset, size_t lenmodule, char *module);
